re #228 refactoring producermessage and stats structs
