+++ src/KafkaW/Producer.h
@@ -2,6 +2,7 @@
 
 #include "BrokerSettings.h"
 #include "ConsumerMessage.h"
+#include "ProducerStats.h"
 #include <atomic>
 #include <functional>
 #include <librdkafka/rdkafkacpp.h>
@@ -9,28 +10,6 @@
 namespace KafkaW {
 
 class ProducerTopic;
-
-struct ProducerMsg {
-  virtual ~ProducerMsg() = default;
-  virtual void deliveryOk();
-  virtual void deliveryError();
-  unsigned char *data;
-  uint32_t size;
-};
-
-struct ProducerStats {
-  std::atomic<uint64_t> produced{0};
-  std::atomic<uint32_t> produce_fail{0};
-  std::atomic<uint32_t> local_queue_full{0};
-  std::atomic<uint64_t> produce_cb{0};
-  std::atomic<uint64_t> produce_cb_fail{0};
-  std::atomic<uint64_t> poll_served{0};
-  std::atomic<uint64_t> msg_too_large{0};
-  std::atomic<uint64_t> produced_bytes{0};
-  std::atomic<uint32_t> out_queue{0};
-  ProducerStats() = default;
-  ProducerStats(ProducerStats const &);
-};
 
 class ProducerInterface {
 public:
@@ -46,7 +25,6 @@ public:
 class Producer : public ProducerInterface {
 public:
   typedef ProducerTopic Topic;
-  typedef ProducerMsg Msg;
   explicit Producer(BrokerSettings ProducerBrokerSettings_);
   Producer(Producer const &) = delete;
   Producer(Producer &&x) noexcept;
