+++ src/KafkaW/ProducerStats.h
@@ -1,0 +1,19 @@
+#pragma once
+#include <atomic>
+namespace KafkaW {
+
+// Provides statistics for all producers
+struct ProducerStats {
+  std::atomic<uint64_t> produced{0};
+  std::atomic<uint32_t> produce_fail{0};
+  std::atomic<uint32_t> local_queue_full{0};
+  std::atomic<uint64_t> produce_cb{0};
+  std::atomic<uint64_t> produce_cb_fail{0};
+  std::atomic<uint64_t> poll_served{0};
+  std::atomic<uint64_t> msg_too_large{0};
+  std::atomic<uint64_t> produced_bytes{0};
+  std::atomic<uint32_t> out_queue{0};
+  ProducerStats() = default;
+  ProducerStats(ProducerStats const &);
+};
+}
